added bench
Former-commit-id: 3d0a8b093a88ee70e796fc6103ec52d0a40d76a7

+++ RustBlackJack/src/table.rs
@@ -108,7 +108,7 @@ impl Table {
     }
 
     fn select_bet(&mut self) {
-        if self.m_truecount as i32 >= 2 {
+        if self.m_truecount >= 2 {
             self.m_players[self.m_currentplayer].m_initialbet =
                 self.m_betsize * (self.m_truecount - 1);
         }
